ai-study-assistant/requirements.txt
flask
google-generativeai
PyPDF2
python-dotenv
werkzeug